refactor and wording
